Rename BeaconRestApi classes and interface
