+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/JsonTypeDefinitionBeaconRestApi.java
@@ -13,9 +13,15 @@
 
 package tech.pegasys.teku.beaconrestapi;
 
-public interface BeaconRestApiType {
+public class JsonTypeDefinitionBeaconRestApi implements BeaconRestApi {
 
-  public void start();
+  @Override
+  public void start() {
+    throw new UnsupportedOperationException("Unsupported start operation");
+  }
 
-  public void stop();
+  @Override
+  public void stop() {
+    throw new UnsupportedOperationException("Unsupported stop operation");
+  }
 }
