+++ services/beaconchain/src/main/java/tech/pegasys/teku/services/beaconchain/BeaconChainController.java
@@ -42,8 +42,8 @@ import tech.pegasys.teku.beacon.sync.SyncService;
 import tech.pegasys.teku.beacon.sync.SyncServiceFactory;
 import tech.pegasys.teku.beacon.sync.events.CoalescingChainHeadChannel;
 import tech.pegasys.teku.beaconrestapi.BeaconRestApi;
-import tech.pegasys.teku.beaconrestapi.BeaconRestApiMigrated;
-import tech.pegasys.teku.beaconrestapi.BeaconRestApiType;
+import tech.pegasys.teku.beaconrestapi.JsonTypeDefinitionBeaconRestApi;
+import tech.pegasys.teku.beaconrestapi.ReflectionBasedBeaconRestApi;
 import tech.pegasys.teku.ethereum.events.SlotEventsChannel;
 import tech.pegasys.teku.ethereum.pow.api.Eth1EventsChannel;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
@@ -189,7 +189,7 @@ public class BeaconChainController extends Service implements BeaconChainControl
   protected volatile BlockImporter blockImporter;
   protected volatile RecentChainData recentChainData;
   protected volatile Eth2P2PNetwork p2pNetwork;
-  protected volatile Optional<BeaconRestApiType> beaconRestAPI = Optional.empty();
+  protected volatile Optional<BeaconRestApi> beaconRestAPI = Optional.empty();
   protected volatile AggregatingAttestationPool attestationPool;
   protected volatile DepositProvider depositProvider;
   protected volatile SyncService syncService;
@@ -256,8 +256,7 @@ public class BeaconChainController extends Service implements BeaconChainControl
     forkChoiceExecutor.start();
     return initialize()
         .thenCompose(
-            (__) ->
-                SafeFuture.fromRunnable(() -> beaconRestAPI.ifPresent(BeaconRestApiType::start)));
+            (__) -> SafeFuture.fromRunnable(() -> beaconRestAPI.ifPresent(BeaconRestApi::start)));
   }
 
   protected void startServices() {
@@ -301,7 +300,7 @@ public class BeaconChainController extends Service implements BeaconChainControl
   protected SafeFuture<?> doStop() {
     LOG.debug("Stopping {}", this.getClass().getSimpleName());
     return SafeFuture.allOf(
-            SafeFuture.fromRunnable(() -> beaconRestAPI.ifPresent(BeaconRestApiType::stop)),
+            SafeFuture.fromRunnable(() -> beaconRestAPI.ifPresent(BeaconRestApi::stop)),
             syncService.stop(),
             blockManager.stop(),
             attestationManager.stop(),
@@ -839,10 +838,10 @@ public class BeaconChainController extends Service implements BeaconChainControl
             .build();
     final Eth1DataProvider eth1DataProvider = new Eth1DataProvider(eth1DataCache, depositProvider);
 
-    final BeaconRestApiType api =
+    final BeaconRestApi api =
         beaconConfig.beaconRestApiConfig().isEnableMigratedRestApi()
-            ? new BeaconRestApiMigrated()
-            : new BeaconRestApi(
+            ? new JsonTypeDefinitionBeaconRestApi()
+            : new ReflectionBasedBeaconRestApi(
                 dataProvider,
                 eth1DataProvider,
                 beaconConfig.beaconRestApiConfig(),
@@ -1171,7 +1170,7 @@ public class BeaconChainController extends Service implements BeaconChainControl
   }
 
   @Override
-  public Optional<BeaconRestApiType> getBeaconRestAPI() {
+  public Optional<BeaconRestApi> getBeaconRestAPI() {
     return beaconRestAPI;
   }
